MCP-544 Fix flaky ConcurrentModificationException in HTTP tests
